[#254] Display Trials' documents in its details page
To make it easier to display the documents data, I added a "presenter", which
can be thought of a View Model. It's useful to add and change the Trials'
attributes for display. Also, being a regular javascript file with no
dependencies to HapiJS or the DB, it's very easy to test.

In the future, a few other places could be refactored into presenters (e.g. the
trials' `contributeDataUrl`).